# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: wolfXmusic API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Search music
  - name: tracks
    description: Track operations
  - name: albums
    description: Album operations
  - name: artists
    description: Artist operations
  - name: favorites
    description: Favorite tracks
  - name: playlists
    description: User playlists

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /music/search:
    get:
      operationId: searchMusic
      tags: [search]
      summary: Search tracks, albums, artists
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [track, album, artist, playlist]
            default: track
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  /music/track/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get track by Spotify ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"

  /music/album/{id}:
    get:
      operationId: getAlbum
      tags: [albums]
      summary: Get album details and tracklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"

  /music/artist/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get artist profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"

  /music/artist/{id}/top-tracks:
    get:
      operationId: getArtistTopTracks
      tags: [artists]
      summary: Get top tracks for an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist top tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /music/artist/{id}/albums:
    get:
      operationId: getArtistAlbums
      tags: [artists]
      summary: Get artist discography
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummary"

  /music/playlist/{id}:
    get:
      operationId: getSpotifyPlaylist
      tags: [search]
      summary: Get Spotify playlist info and tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playlist tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotifyPlaylist"

  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get all favorited tracks
      responses:
        "200":
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a track to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Favorite added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"

  /favorites/{trackId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a track from favorites
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /playlists:
    get:
      operationId: getPlaylists
      tags: [playlists]
      summary: List all user playlists
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"

  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get playlist with tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithTracks"
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update playlist metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /playlists/{id}/tracks:
    post:
      operationId: addTrackToPlaylist
      tags: [playlists]
      summary: Add a track to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistTrackInput"
      responses:
        "201":
          description: Track added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /playlists/{id}/tracks/{trackId}:
    delete:
      operationId: removeTrackFromPlaylist
      tags: [playlists]
      summary: Remove a track from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Track:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        artists:
          type: array
          items:
            type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        explicit:
          type: boolean
        preview_url:
          type: ["string", "null"]
        url:
          type: string
      required:
        - id
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - explicit
        - url

    AlbumSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        release_date:
          type: string
        total_tracks:
          type: integer
        thumbnail:
          type: string
        artist:
          type: string
        url:
          type: string
      required:
        - id
        - name
        - thumbnail
        - artist

    Album:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artist:
          type: string
        thumbnail:
          type: string
        release_date:
          type: string
        total_tracks:
          type: integer
        type:
          type: string
        url:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
      required:
        - id
        - name
        - artist
        - thumbnail
        - tracks

    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        thumbnail:
          type: string
        followers:
          type: integer
        genres:
          type: array
          items:
            type: string
        verified:
          type: boolean
        url:
          type: string
      required:
        - id
        - name
        - thumbnail

    SearchTrackResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        explicit:
          type: boolean
        url:
          type: string
      required:
        - id
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - explicit
        - url

    SearchArtistResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        thumbnail:
          type: string
        url:
          type: string
      required:
        - id
        - name
        - thumbnail

    SearchAlbumResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artist:
          type: string
        thumbnail:
          type: string
        release_date:
          type: string
        url:
          type: string
      required:
        - id
        - name
        - artist
        - thumbnail

    SearchResults:
      type: object
      properties:
        query:
          type: string
        type:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            type: object
      required:
        - query
        - type
        - total
        - results

    SpotifyPlaylist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        total_tracks:
          type: integer
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
      required:
        - id
        - name
        - tracks

    Favorite:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        previewUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: string
        createdAt:
          type: string
      required:
        - id
        - trackId
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - spotifyUrl
        - createdAt

    FavoriteInput:
      type: object
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        previewUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: string
      required:
        - trackId
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - spotifyUrl

    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        trackCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - trackCount
        - createdAt
        - updatedAt

    PlaylistInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name

    PlaylistUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: string

    PlaylistTrack:
      type: object
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        previewUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: string
        position:
          type: integer
        addedAt:
          type: string
      required:
        - id
        - playlistId
        - trackId
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - spotifyUrl
        - position
        - addedAt

    PlaylistWithTracks:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        trackCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistTrack"
      required:
        - id
        - name
        - trackCount
        - createdAt
        - updatedAt
        - tracks

    PlaylistTrackInput:
      type: object
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        duration_ms:
          type: integer
        previewUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: string
      required:
        - trackId
        - title
        - artist
        - album
        - thumbnail
        - duration
        - duration_ms
        - spotifyUrl