# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from './track';

export interface Album {
  id: string;
  name: string;
  artist: string;
  thumbnail: string;
  release_date?: string;
  total_tracks?: number;
  type?: string;
  url?: string;
  tracks: Track[];
}