# lib/api-zod/src/generated/types/albumSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface AlbumSummary {
  id: string;
  name: string;
  type?: string;
  release_date?: string;
  total_tracks?: number;
  thumbnail: string;
  artist: string;
  url?: string;
}