# lib/api-zod/src/generated/types/artist.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface Artist {
  id: string;
  name: string;
  thumbnail: string;
  followers?: number;
  genres?: string[];
  verified?: boolean;
  url?: string;
}