# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  trackId: string;
  title: string;
  artist: string;
  album: string;
  thumbnail: string;
  duration: string;
  duration_ms: number;
  /** @nullable */
  previewUrl?: string | null;
  spotifyUrl: string;
  createdAt: string;
}