# lib/api-zod/src/generated/types/playlistUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistUpdate {
  name?: string;
  description?: string;
  thumbnail?: string;
}