# lib/api-zod/src/generated/types/playlistWithTracks.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistTrack } from './playlistTrack';

export interface PlaylistWithTracks {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  trackCount: number;
  createdAt: string;
  updatedAt: string;
  tracks: PlaylistTrack[];
}