# lib/api-zod/src/generated/types/searchAlbumResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchAlbumResult {
  id: string;
  name: string;
  artist: string;
  thumbnail: string;
  release_date?: string;
  url?: string;
}