# lib/api-zod/src/generated/types/searchArtistResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * wolfXmusic API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchArtistResult {
  id: string;
  name: string;
  thumbnail: string;
  url?: string;
}